+++ frontend/src/admin/PageHeader.tsx
@@ -1,0 +1,18 @@
+import { Typography } from '@mui/material';
+
+function PageHeader() {
+  // const user = {
+  //   name: 'Catherine Pike',
+  //   avatar: '/static/images/avatars/1.jpg'
+  // };
+
+  return (
+    <>
+      <Typography variant="h3" component="h3" gutterBottom>
+        Real-Time Billboard Auction
+      </Typography>
+    </>
+  );
+}
+
+export default PageHeader;
