--- frontend/src/content/applications/Users/settings/PageHeader.tsx
@@ -1,21 +1,0 @@
-import { Typography } from '@mui/material';
-
-function PageHeader() {
-  const user = {
-    name: 'Catherine Pike',
-    avatar: '/static/images/avatars/1.jpg'
-  };
-
-  return (
-    <>
-      <Typography variant="h3" component="h3" gutterBottom>
-        User Settings
-      </Typography>
-      <Typography variant="subtitle2">
-        {user.name}, this could be your user settings panel.
-      </Typography>
-    </>
-  );
-}
-
-export default PageHeader;
