+++ frontend/src/layouts/SidebarLayout/Header/index.tsx
@@ -50,10 +50,10 @@ function Header() {
       }}
     >
       <Stack direction="row" divider={<Divider orientation="vertical" flexItem />} alignItems="center" spacing={2}>
-        <HeaderMenu />
+        {/* <HeaderMenu /> */}
       </Stack>
       <Box display="flex" alignItems="center">
-        <HeaderButtons />
+        {/* <HeaderButtons /> */}
         <Web3Login />
         <Box
           component="span"
